+++ src/lib/ares_private.h
@@ -52,11 +52,13 @@
 #include "ares_htable_szvp.h"
 #include "ares_htable_asvp.h"
 #include "ares_htable_vpvp.h"
+#include "ares_htable_vpstr.h"
 #include "record/ares_dns_multistring.h"
 #include "ares_buf.h"
 #include "record/ares_dns_private.h"
 #include "util/ares_iface_ips.h"
 #include "util/ares_threads.h"
+#include "ares_socket.h"
 #include "ares_conn.h"
 #include "ares_str.h"
 #include "str/ares_strsplit.h"
