+++ src/lib/dsa/ares_htable_vpstr.c
@@ -1,6 +1,6 @@
 /* MIT License
  *
- * Copyright (c) 2023 Brad House
+ * Copyright (c) 2024 Brad House
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -25,23 +25,22 @@
  */
 #include "ares_private.h"
 #include "ares_htable.h"
-#include "ares_htable_strvp.h"
+#include "ares_htable_vpstr.h"
 
-struct ares_htable_strvp {
-  ares_htable_strvp_val_free_t free_val;
-  ares_htable_t               *hash;
+struct ares_htable_vpstr {
+  ares_htable_t *hash;
 };
 
 typedef struct {
-  char                *key;
-  void                *val;
-  ares_htable_strvp_t *parent;
-} ares_htable_strvp_bucket_t;
+  void                *key;
+  char                *val;
+  ares_htable_vpstr_t *parent;
+} ares_htable_vpstr_bucket_t;
 
-void ares_htable_strvp_destroy(ares_htable_strvp_t *htable)
+void ares_htable_vpstr_destroy(ares_htable_vpstr_t *htable)
 {
   if (htable == NULL) {
-    return;
+    return; /* LCOV_EXCL_LINE: DefensiveCoding */
   }
 
   ares_htable_destroy(htable->hash);
@@ -50,64 +49,63 @@ void ares_htable_strvp_destroy(ares_htable_strvp_t *htable)
 
 static unsigned int hash_func(const void *key, unsigned int seed)
 {
-  const char *arg = key;
-  return ares_htable_hash_FNV1a_casecmp((const unsigned char *)arg,
-                                        ares_strlen(arg), seed);
+  return ares_htable_hash_FNV1a((const unsigned char *)&key, sizeof(key), seed);
 }
 
 static const void *bucket_key(const void *bucket)
 {
-  const ares_htable_strvp_bucket_t *arg = bucket;
+  const ares_htable_vpstr_bucket_t *arg = bucket;
   return arg->key;
 }
 
 static void bucket_free(void *bucket)
 {
-  ares_htable_strvp_bucket_t *arg = bucket;
+  ares_htable_vpstr_bucket_t *arg = bucket;
 
-  if (arg->parent->free_val) {
-    arg->parent->free_val(arg->val);
-  }
-  ares_free(arg->key);
+  ares_free(arg->val);
+
   ares_free(arg);
 }
 
 static ares_bool_t key_eq(const void *key1, const void *key2)
 {
-  return ares_strcaseeq(key1, key2);
+  if (key1 == key2) {
+    return ARES_TRUE;
+  }
+
+  return ARES_FALSE;
 }
 
-ares_htable_strvp_t *
-  ares_htable_strvp_create(ares_htable_strvp_val_free_t val_free)
+ares_htable_vpstr_t *ares_htable_vpstr_create(void)
 {
-  ares_htable_strvp_t *htable = ares_malloc(sizeof(*htable));
+  ares_htable_vpstr_t *htable = ares_malloc(sizeof(*htable));
   if (htable == NULL) {
-    goto fail;
+    goto fail; /* LCOV_EXCL_LINE: OutOfMemory */
   }
 
   htable->hash = ares_htable_create(hash_func, bucket_key, bucket_free, key_eq);
   if (htable->hash == NULL) {
-    goto fail;
+    goto fail; /* LCOV_EXCL_LINE: OutOfMemory */
   }
-
-  htable->free_val = val_free;
 
   return htable;
 
+/* LCOV_EXCL_START: OutOfMemory */
 fail:
   if (htable) {
     ares_htable_destroy(htable->hash);
     ares_free(htable);
   }
   return NULL;
+  /* LCOV_EXCL_STOP */
 }
 
-ares_bool_t ares_htable_strvp_insert(ares_htable_strvp_t *htable,
-                                     const char *key, void *val)
+ares_bool_t ares_htable_vpstr_insert(ares_htable_vpstr_t *htable, void *key,
+                                     const char *val)
 {
-  ares_htable_strvp_bucket_t *bucket = NULL;
+  ares_htable_vpstr_bucket_t *bucket = NULL;
 
-  if (htable == NULL || key == NULL) {
+  if (htable == NULL) {
     goto fail;
   }
 
@@ -117,11 +115,11 @@ ares_bool_t ares_htable_strvp_insert(ares_htable_strvp_t *htable,
   }
 
   bucket->parent = htable;
-  bucket->key    = ares_strdup(key);
-  if (bucket->key == NULL) {
+  bucket->key    = key;
+  bucket->val    = ares_strdup(val);
+  if (bucket->val == NULL) {
     goto fail;
   }
-  bucket->val = val;
 
   if (!ares_htable_insert(htable->hash, bucket)) {
     goto fail;
@@ -131,22 +129,22 @@ ares_bool_t ares_htable_strvp_insert(ares_htable_strvp_t *htable,
 
 fail:
   if (bucket) {
-    ares_free(bucket->key);
+    ares_free(bucket->val);
     ares_free(bucket);
   }
   return ARES_FALSE;
 }
 
-ares_bool_t ares_htable_strvp_get(const ares_htable_strvp_t *htable,
-                                  const char *key, void **val)
+ares_bool_t ares_htable_vpstr_get(const ares_htable_vpstr_t *htable,
+                                  const void *key, const char **val)
 {
-  ares_htable_strvp_bucket_t *bucket = NULL;
+  ares_htable_vpstr_bucket_t *bucket = NULL;
 
   if (val) {
     *val = NULL;
   }
 
-  if (htable == NULL || key == NULL) {
+  if (htable == NULL) {
     return ARES_FALSE;
   }
 
@@ -161,16 +159,16 @@ ares_bool_t ares_htable_strvp_get(const ares_htable_strvp_t *htable,
   return ARES_TRUE;
 }
 
-void *ares_htable_strvp_get_direct(const ares_htable_strvp_t *htable,
-                                   const char                *key)
+const char *ares_htable_vpstr_get_direct(const ares_htable_vpstr_t *htable,
+                                         const void                *key)
 {
-  void *val = NULL;
-  ares_htable_strvp_get(htable, key, &val);
+  const char *val = NULL;
+  ares_htable_vpstr_get(htable, key, &val);
   return val;
 }
 
-ares_bool_t ares_htable_strvp_remove(ares_htable_strvp_t *htable,
-                                     const char          *key)
+ares_bool_t ares_htable_vpstr_remove(ares_htable_vpstr_t *htable,
+                                     const void          *key)
 {
   if (htable == NULL) {
     return ARES_FALSE;
@@ -179,29 +177,7 @@ ares_bool_t ares_htable_strvp_remove(ares_htable_strvp_t *htable,
   return ares_htable_remove(htable->hash, key);
 }
 
-void *ares_htable_strvp_claim(ares_htable_strvp_t *htable, const char *key)
-{
-  ares_htable_strvp_bucket_t *bucket = NULL;
-  void                       *val;
-
-  if (htable == NULL || key == NULL) {
-    return NULL;
-  }
-
-  bucket = ares_htable_get(htable->hash, key);
-  if (bucket == NULL) {
-    return NULL;
-  }
-
-  /* Unassociate value from bucket */
-  val         = bucket->val;
-  bucket->val = NULL;
-
-  ares_htable_strvp_remove(htable, key);
-  return val;
-}
-
-size_t ares_htable_strvp_num_keys(const ares_htable_strvp_t *htable)
+size_t ares_htable_vpstr_num_keys(const ares_htable_vpstr_t *htable)
 {
   if (htable == NULL) {
     return 0;
