+++ src/lib/include/ares_htable_vpstr.h
@@ -1,6 +1,6 @@
 /* MIT License
  *
- * Copyright (c) 2023 Brad House
+ * Copyright (c) 2024 Brad House
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -23,14 +23,14 @@
  *
  * SPDX-License-Identifier: MIT
  */
-#ifndef __ARES__HTABLE_STRVP_H
-#define __ARES__HTABLE_STRVP_H
+#ifndef __ARES__HTABLE_VPSTR_H
+#define __ARES__HTABLE_VPSTR_H
 
-/*! \addtogroup ares_htable_strvp HashTable with string Key and void pointer
- * Value
+/*! \addtogroup ares_htable_vpstr HashTable with void pointer Key and string
+ *  value
  *
  * This data structure wraps the base ares_htable data structure in order to
- * split the key and value data types as string and void pointer, respectively.
+ * split the key and value data types as void pointer and string, respectively.
  *
  * Average time complexity:
  *  - Insert: O(1)
@@ -40,41 +40,32 @@
  * @{
  */
 
-struct ares_htable_strvp;
+struct ares_htable_vpstr;
 
-/*! Opaque data type for size_t key, void pointer hash table implementation */
-typedef struct ares_htable_strvp ares_htable_strvp_t;
-
-/*! Callback to free value stored in hashtable
- *
- *  \param[in] val  user-supplied value
- */
-typedef void (*ares_htable_strvp_val_free_t)(void *val);
+/*! Opaque data type for void pointer key, string value hash table
+ * implementation */
+typedef struct ares_htable_vpstr ares_htable_vpstr_t;
 
 /*! Destroy hashtable
  *
  *  \param[in] htable  Initialized hashtable
  */
-CARES_EXTERN void ares_htable_strvp_destroy(ares_htable_strvp_t *htable);
+CARES_EXTERN void ares_htable_vpstr_destroy(ares_htable_vpstr_t *htable);
 
-/*! Create string, void pointer value hash table
+/*! Create void pointer key, string value hash table
  *
- *  \param[in] val_free  Optional. Call back to free user-supplied value.  If
- *                       NULL it is expected the caller will clean up any user
- *                       supplied values.
  */
-CARES_EXTERN ares_htable_strvp_t *
-  ares_htable_strvp_create(ares_htable_strvp_val_free_t val_free);
+CARES_EXTERN ares_htable_vpstr_t *ares_htable_vpstr_create(void);
 
 /*! Insert key/value into hash table
  *
  *  \param[in] htable Initialized hash table
  *  \param[in] key    key to associate with value
- *  \param[in] val    value to store (takes ownership). May be NULL.
+ *  \param[in] val    value to store (duplicates).
  *  \return ARES_TRUE on success, ARES_FALSE on failure or out of memory
  */
-CARES_EXTERN ares_bool_t ares_htable_strvp_insert(ares_htable_strvp_t *htable,
-                                                  const char *key, void *val);
+CARES_EXTERN ares_bool_t ares_htable_vpstr_insert(ares_htable_vpstr_t *htable,
+                                                  void *key, const char *val);
 
 /*! Retrieve value from hashtable based on key
  *
@@ -83,20 +74,20 @@ CARES_EXTERN ares_bool_t ares_htable_strvp_insert(ares_htable_strvp_t *htable,
  *  \param[out] val     Optional.  Pointer to store value.
  *  \return ARES_TRUE on success, ARES_FALSE on failure
  */
-CARES_EXTERN ares_bool_t ares_htable_strvp_get(
-  const ares_htable_strvp_t *htable, const char *key, void **val);
+CARES_EXTERN ares_bool_t ares_htable_vpstr_get(
+  const ares_htable_vpstr_t *htable, const void *key, const char **val);
 
 /*! Retrieve value from hashtable directly as return value.  Caveat to this
- *  function over ares_htable_strvp_get() is that if a NULL value is stored
+ *  function over ares_htable_vpstr_get() is that if a NULL value is stored
  *  you cannot determine if the key is not found or the value is NULL.
  *
  *  \param[in] htable  Initialized hash table
  *  \param[in] key     key to use to search
  *  \return value associated with key in hashtable or NULL
  */
-CARES_EXTERN void *
-  ares_htable_strvp_get_direct(const ares_htable_strvp_t *htable,
-                               const char                *key);
+CARES_EXTERN const char *
+  ares_htable_vpstr_get_direct(const ares_htable_vpstr_t *htable,
+                               const void                *key);
 
 /*! Remove a value from the hashtable by key
  *
@@ -104,18 +95,8 @@ CARES_EXTERN void *
  *  \param[in] key     key to use to search
  *  \return ARES_TRUE if found, ARES_FALSE if not
  */
-CARES_EXTERN ares_bool_t ares_htable_strvp_remove(ares_htable_strvp_t *htable,
-                                                  const char          *key);
-
-/*! Remove the value from the hashtable, and return the value instead of
- *  calling the val_free passed to ares_htable_strvp_create().
- *
- *  \param[in] htable  Initialized hash table
- *  \param[in] key     key to use to search
- *  \return value in hashtable or NULL on error
- */
-CARES_EXTERN void       *ares_htable_strvp_claim(ares_htable_strvp_t *htable,
-                                                 const char          *key);
+CARES_EXTERN ares_bool_t ares_htable_vpstr_remove(ares_htable_vpstr_t *htable,
+                                                  const void          *key);
 
 /*! Retrieve the number of keys stored in the hash table
  *
@@ -123,8 +104,8 @@ CARES_EXTERN void       *ares_htable_strvp_claim(ares_htable_strvp_t *htable,
  *  \return count
  */
 CARES_EXTERN size_t
-  ares_htable_strvp_num_keys(const ares_htable_strvp_t *htable);
+  ares_htable_vpstr_num_keys(const ares_htable_vpstr_t *htable);
 
 /*! @} */
 
-#endif /* __ARES__HTABLE_STRVP_H */
+#endif /* __ARES__HTABLE_VPSTR_H */
